+++ core/src/main/java/me/rhin/openciv/game/city/building/type/Library.java
@@ -18,7 +18,7 @@ public class Library extends Building implements SpecialistContainer {
 	}
 
 	@Override
-	public int getProductionCost() {
+	public float getProductionCost() {
 		return 75;
 	}
