+++ core/src/main/java/me/rhin/openciv/game/city/building/type/WaterMill.java
@@ -22,7 +22,7 @@ public class WaterMill extends Building {
 	}
 
 	@Override
-	public int getProductionCost() {
+	public float getProductionCost() {
 		return 75;
 	}
