+++ server/src/main/java/me/rhin/openciv/server/game/city/building/type/Library.java
@@ -47,7 +47,7 @@ public class Library extends Building implements CityGrowthListener, CityStarveL
 	}
 
 	@Override
-	public int getProductionCost() {
+	public float getProductionCost() {
 		return 75;
 	}
